Rename _online_tools to _online_helper
