+++ pyriodicity/tools/__init__.py
@@ -1,4 +1,4 @@
-from ._online_tools import OnlineHelper
+from ._online_helper import OnlineHelper
 from ._tools import acf, apply_window, power_threshold, to_1d_array
 
 __all__ = ["OnlineHelper", "acf", "apply_window", "power_threshold", "to_1d_array"]
